+++ docs/about/changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.1.2 - 2025-10-19
+
+### Fixed
+
+- Use direct file system path for static files to fix the issue where the network graph static assets were not copied to the site directory.
+
+### Added
+
+- Added `MANIFEST.in` file to include the network graph static assets in the package.
+
 ## 2.1.1 - 2025-10-19
 
 ### Fixed